refactor to toggle between views
